Summary: improvements and additions. Details: 	1. environment/defaults/bin/buildApidoc.sh: improvements. 	2. environment/defaults/completions/dcr: support for 2 or 4 space indentations in yml files; support for listing dcr's own options. 	3. environment/xcustoms/crpi.base/atom/config.cson: udpates from atom. 	4. environment/xcustoms/crpi.base/atom/snippets.cson: same as point 3. 	5. environment/xcustoms/crpi.base/bin/add: updated to better handle apidoc files. 	6. environment/xcustoms/crpi.work/.aliases_custom: added aliases. 	7. Deleted file environment/xcustoms/crpi.work/.kube/.gitignore. 	8. environment/xcustoms/crpi.work/.synergy.conf: updated some IP addresses. 	9. environment/xcustoms/crpi.work/.userenv_custom: some updates. 	10. Added file environment/xcustoms/crpi.work/CRON/cleanupDocker.sh. 	11. Added file environment/xcustoms/crpi.work/CRON/cron.tab. 	12. Added file environment/xcustoms/crpi.work/bin/build. 	13. environment/xcustoms/crpi.work/bin/compactizeJson: support for recognizing settings and scopes json. 	14. environment/xcustoms/crpi.work/bin/dcr: made bash the default command. 	15. environment/xcustoms/crpi.work/bin/tagRelease.sh: added code to insert a new section in the CHANGELOG file. 	16. environment/xcustoms/crpi.work/user.config.install: updated according to points 10 to 12.

+++ environment/defaults/bin/buildApidoc.sh
@@ -37,18 +37,31 @@ function buildFile()
     local lInputFile=${1}
     local lOutputFile=${2}
     
-    # Only build if the output file is older than the input file
-    if [[ ${force} == true ]] || [[ ${lOutputFile} -ot ${lInputFile} ]]; then
-	versionFile=.version
-	if [[ -f ${versionFile} ]]; then
-	    version=$(cat ${versionFile})
+
+    local lVersionFile=.version
+    # Only build if force
+    if [[ ${force} == true ]] || \
+	   # OR the output file doesn't exist
+	   [[ ! -f ${lOutputFile} ]] || \
+	   # OR the output file is older than the input file
+	   [[ ${lOutputFile} -ot ${lInputFile} ]] || \
+	   # OR the output file is older than the version file
+           [[ ${lOutputFile} -ot ${lVersionFile} ]]; then
+	echo Building ${lInputFile} to ${lOutputFile}
+
+	# If version file found, use it
+	if [[ -f ${lVersionFile} ]]; then
+	    version=$(cat ${lVersionFile})
 	    tmpFile=$(mktemp -p $(dirname ${lInputFile}))
 	    cat ${lInputFile} | sed "s/__VERSION__/${version}/" > ${tmpFile}
 	    lInputFile=${tmpFile}
 	fi
 
 	aglio -i ${lInputFile} --theme-template triple -o ${lOutputFile}
-	test -f ${tmpFile} && rm ${tmpFile}
+
+	if [[ -n ${tmpFile} ]]; then
+	    test -f ${tmpFile} && rm ${tmpFile}
+	fi
     fi
 }
 
@@ -112,7 +125,7 @@ fi
 
 inputFile="${1}"
 if [[ -z ${inputFile} ]]; then
-    inputFiles=( $(find doc -name '*.apib') )
+    inputFiles=( $(find doc -name 'apidoc*.apib') )
     if [[ -z ${#inputFiles[@]} ]]; then
         echo "Could not find an apib file in doc directory"
         exit 0
@@ -131,6 +144,5 @@ fi
 
 for inputFile in ${inputFiles[@]}; do
     outputFile=${inputFile%.*}'.html'
-    echo Building ${inputFile} to ${outputFile}
     buildFile ${inputFile} ${outputFile}
 done
